git-svn-id: https://svn.syssoft.uni-trier.de/svn/prisma/DeltaWorks/include@2288 87af6b1e-9b34-e211-997a-bc305bce1b9a

+++ string/str_class.cpp
@@ -407,7 +407,7 @@ void ErrMessageW(const StringW&line)
 
 static char* pointerToHex(const void*pointer, int min_len, char*end, char*first)
 {
-	unsigned long value = (unsigned long)pointer;
+	size_t value = (size_t)pointer;
 	while (value && end != first)
 	{
 		BYTE c = value & (0xF);
